ai-service/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
torch==2.1.2
numpy==1.26.2
scipy==1.11.4

# HTTP and API
requests==2.31.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Logging and monitoring  
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
isort==5.13.2